+++ CHANGELOG.md
@@ -9,6 +9,8 @@ Unreleased
   functionality
 - Introduced `Framebuffer` type
 - Made various `enum` types non-exhaustive
+- Added `sys::VertexBufferUsage` enum and adjusted
+  `sys::Context::set_vertex_buffer_data` to require an instance
 - Renamed `PrimitiveTy` trait to `BuiltinType`
 - Renamed `DrawType` enum to `Primitive`